refactor: Player.kt 클래스 sealed class 중첩 클래스로 이동
