+++ src/main/kotlin/blackjack/domain/Participants.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.domain.Participant.Player
+
 data class Participants(private val players: List<Player>) {
     constructor(participants: Participants, dealer: Dealer) : this(listOf(dealer) + participants.players)
