+++ src/main/kotlin/blackjack/domain/PlayerOutcomes.kt
@@ -1,5 +1,7 @@
 package blackjack.domain
 
+import blackjack.domain.Participant.Player
+
 data class PlayerOutcomes(val player: Player, val results: Result) {
     companion object {
         fun from(
